Move sort comparator to own class, rename search thread method
